vertex_trigger_function/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.25.0